解决cookies跨域问题

+++ moj-backend-microservice/moj-backend-common/src/main/java/com/mar/mojbackendcommon/common/PageRequest.java
@@ -1,7 +1,7 @@
 package com.mar.mojbackendcommon.common;
 
-import lombok.Data;
 import com.mar.mojbackendcommon.constant.CommonConstant;
+import lombok.Data;
 
 /**
  * 分页请求
@@ -13,12 +13,12 @@ public class PageRequest {
     /**
      * 当前页号
      */
-    private int current = 1;
+    private long current = 1;
 
     /**
      * 页面大小
      */
-    private int pageSize = 10;
+    private long pageSize = 10;
 
     /**
      * 排序字段
